+++ packages/webview_flutter/CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 0.1.0
 
 * Initial release
+
+## 0.2.0
+* Update Dart and Flutter SDK constraints
+* Update Flutter and Samsung copyright information
+* Update webview_flutter_tizen to use platform view interface
+* Update example and integration_test
+* Update webivew_flutter to 2.0.4